+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Not released
 
-- isRemoteCalculationSupported: dynamic aggregation of points to grid/h3 should be remote [#917](https://github.com/CartoDB/carto-react/pull/917)
+- isRemoteCalculationSupported: remove backwards compatbility check for spatial indices [#917](https://github.com/CartoDB/carto-react/pull/917)
 
 ## 3.0.0